pyproject.toml
[project]
name = "python-fastapi-uv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.118.0",
    "mssql-python>=0.13.0",
    "python-dotenv>=1.1.1",
]
